move accountlist to component packes
