+++ app/src/main/AndroidManifest.xml
@@ -111,7 +111,7 @@
         <activity android:name=".EditProfileActivity" />
         <activity android:name=".components.preference.PreferencesActivity" />
         <activity android:name=".StatusListActivity" />
-        <activity android:name=".AccountListActivity" />
+        <activity android:name=".components.accountlist.AccountListActivity" />
         <activity android:name=".AboutActivity" />
         <activity android:name=".TabPreferenceActivity" />
         <activity
