+++ app/src/main/java/com/keylesspalace/tusky/components/accountlist/AccountListFragment.kt
@@ -13,7 +13,7 @@
  * You should have received a copy of the GNU General Public License along with Tusky; if not,
  * see <http://www.gnu.org/licenses>. */
 
-package com.keylesspalace.tusky.fragment
+package com.keylesspalace.tusky.components.accountlist
 
 import android.os.Bundle
 import android.util.Log
@@ -30,16 +30,16 @@ import androidx.recyclerview.widget.SimpleItemAnimator
 import autodispose2.androidx.lifecycle.AndroidLifecycleScopeProvider.from
 import autodispose2.autoDispose
 import com.google.android.material.snackbar.Snackbar
-import com.keylesspalace.tusky.AccountListActivity.Type
+import com.keylesspalace.tusky.components.accountlist.AccountListActivity.Type
 import com.keylesspalace.tusky.BaseActivity
 import com.keylesspalace.tusky.R
 import com.keylesspalace.tusky.adapter.AccountAdapter
-import com.keylesspalace.tusky.adapter.BlocksAdapter
-import com.keylesspalace.tusky.adapter.FollowAdapter
-import com.keylesspalace.tusky.adapter.FollowRequestsAdapter
-import com.keylesspalace.tusky.adapter.FollowRequestsHeaderAdapter
-import com.keylesspalace.tusky.adapter.MutesAdapter
+import com.keylesspalace.tusky.components.accountlist.adapter.FollowRequestsHeaderAdapter
 import com.keylesspalace.tusky.components.account.AccountActivity
+import com.keylesspalace.tusky.components.accountlist.adapter.BlocksAdapter
+import com.keylesspalace.tusky.components.accountlist.adapter.FollowAdapter
+import com.keylesspalace.tusky.components.accountlist.adapter.FollowRequestsAdapter
+import com.keylesspalace.tusky.components.accountlist.adapter.MutesAdapter
 import com.keylesspalace.tusky.databinding.FragmentAccountListBinding
 import com.keylesspalace.tusky.db.AccountManager
 import com.keylesspalace.tusky.di.Injectable
@@ -100,7 +100,9 @@ class AccountListFragment : Fragment(R.layout.fragment_account_list), AccountAct
             Type.BLOCKS -> BlocksAdapter(this, animateAvatar, animateEmojis, showBotOverlay)
             Type.MUTES -> MutesAdapter(this, animateAvatar, animateEmojis, showBotOverlay)
             Type.FOLLOW_REQUESTS -> {
-                val headerAdapter = FollowRequestsHeaderAdapter(accountManager.activeAccount!!.domain, arguments?.getBoolean(ARG_ACCOUNT_LOCKED) == true)
+                val headerAdapter = FollowRequestsHeaderAdapter(accountManager.activeAccount!!.domain, arguments?.getBoolean(
+                    ARG_ACCOUNT_LOCKED
+                ) == true)
                 val followRequestsAdapter = FollowRequestsAdapter(this, animateAvatar, animateEmojis, showBotOverlay)
                 binding.recyclerView.adapter = ConcatAdapter(headerAdapter, followRequestsAdapter)
                 followRequestsAdapter
