+++ app/src/main/java/com/keylesspalace/tusky/components/accountlist/adapter/FollowAdapter.kt
@@ -12,10 +12,12 @@
  *
  * You should have received a copy of the GNU General Public License along with Tusky; if not,
  * see <http://www.gnu.org/licenses>. */
-package com.keylesspalace.tusky.adapter
+package com.keylesspalace.tusky.components.accountlist.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
+import com.keylesspalace.tusky.adapter.AccountAdapter
+import com.keylesspalace.tusky.adapter.AccountViewHolder
 import com.keylesspalace.tusky.databinding.ItemAccountBinding
 import com.keylesspalace.tusky.interfaces.AccountActionListener
