+++ app/src/main/java/com/keylesspalace/tusky/components/accountlist/adapter/MutesAdapter.kt
@@ -1,9 +1,10 @@
-package com.keylesspalace.tusky.adapter
+package com.keylesspalace.tusky.components.accountlist.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.core.view.ViewCompat
 import com.keylesspalace.tusky.R
+import com.keylesspalace.tusky.adapter.AccountAdapter
 import com.keylesspalace.tusky.databinding.ItemMutedUserBinding
 import com.keylesspalace.tusky.interfaces.AccountActionListener
 import com.keylesspalace.tusky.util.BindingHolder
